add: adding flyway in the project and adding some sql scripts.

+++ ecocoleta-api/src/main/java/com/github/jvtopsilva090/ecocoleta/ecocoletaapi/dto/CollectionPointCreateDto.java
@@ -1,11 +1,12 @@
 package com.github.jvtopsilva090.ecocoleta.ecocoletaapi.dto;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 public record CollectionPointCreateDto(
     String name,
     String formattedAddress,
-    Double latitude,
-    Double longitude,
-    List<Long> residueIds
+    BigDecimal latitude,
+    BigDecimal longitude,
+    List<Integer> residueIds
 ) {}
