+++ ecocoleta-api/src/main/java/com/github/jvtopsilva090/ecocoleta/ecocoletaapi/dto/CollectionPointOutDto.java
@@ -2,14 +2,15 @@ package com.github.jvtopsilva090.ecocoleta.ecocoletaapi.dto;
 
 import com.github.jvtopsilva090.ecocoleta.ecocoletaapi.entity.CollectionPoint;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 public record CollectionPointOutDto(
-    Long id,
+    Integer id,
     String name,
     String formattedAddress,
-    Double latitude,
-    Double longitude,
+    BigDecimal latitude,
+    BigDecimal longitude,
     List<Object> residuesType
 ) {
     public CollectionPointOutDto(CollectionPoint collectionPoint) {
