+++ ecocoleta-api/src/main/java/com/github/jvtopsilva090/ecocoleta/ecocoletaapi/entity/CollectionPointResiduesId.java
@@ -13,8 +13,8 @@ import lombok.NoArgsConstructor;
 public class CollectionPointResiduesId {
 
     @Column(name = "collection_point_id", nullable = false)
-    private Long collectionPointId;
+    private Integer collectionPointId;
 
     @Column(name = "residue_id", nullable = false)
-    private Long residueId;
+    private Integer residueId;
 }
