+++ ecocoleta-api/src/main/java/com/github/jvtopsilva090/ecocoleta/ecocoletaapi/repository/ResidueRepository.java
@@ -3,5 +3,5 @@ package com.github.jvtopsilva090.ecocoleta.ecocoletaapi.repository;
 import com.github.jvtopsilva090.ecocoleta.ecocoletaapi.entity.Residue;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ResidueRepository extends JpaRepository<Residue, Long> {
+public interface ResidueRepository extends JpaRepository<Residue, Integer> {
 }
